create modul models containing all models
